+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 --- 
 # Beta - published
 
+## [0.2.4] - 2026-08-14
+- [Module] fixed module running not affected Unity runtime phase
+- [Module] add event to listen for all module phase completed
+- [Sample] all module phase sample and summary
+
 ## [0.2.3] - 2026-08-13
 - [Quit Pipeline] fixed build issue - hanged on safe quit process on empty queue 
 
@@ -23,5 +28,5 @@ All notable changes to this project will be documented in this file.
 dev optimize and bug fix
 
 ## [0.1.0] - 2026-08-11
-Repo created, project in dev level, not published.
+Repo created, project in dev level, not published.   
 For detail please visit and bookmark our [GitBook](https://aceland-workshop.gitbook.io/aceland-unity-packages/)